Use "--interactive=false" to avoid asking for sudo password
